Added capability to perform local processing

+++ scripts/make_ref_speeds.py
@@ -176,6 +176,7 @@ if __name__ == "__main__":
   parser.add_argument('--level', type=int, help='The level to target', required=True)
   parser.add_argument('--tile-id', type=int, help='The tile id to target', required=True)
   parser.add_argument('--no-separate-subtiles', help='If present all subtiles will be in the same tile', action='store_true')
+  parser.add_argument('--local', help='Enable local file processing', action='store_true')
   parser.add_argument('--verbose', '-v', help='Turn on verbose output i.e. DEBUG level logging', action='store_true')
 
   # parse the arguments
@@ -192,50 +193,54 @@ if __name__ == "__main__":
   else:
     log.basicConfig(format='%(asctime)s - %(levelname)s - %(message)s', stream=sys.stdout)
 
-################################################################################
-
+  ################################################################################
   # download the speed tiles from aws and decompress
-  spdFileNames = []
-  directory = "ref_working_dir/"
-  shutil.rmtree(directory, ignore_errors=True)
-  tile_hierarchy = TileHierarchy()
-  key_prefix = args.year + "/"
+  if not args.local:
+    log.debug('Download the speed tiles from AWS and decompress...')
+    spdFileNames = []
+    directory = "ref_working_dir/"
+    shutil.rmtree(directory, ignore_errors=True)
+    tile_hierarchy = TileHierarchy()
+    key_prefix = args.year + "/"
 
-  s3 = boto3.client('s3')
+    s3 = boto3.client('s3')
 
-  weeks_per_year = 52
-  week = 0
-  while ( week < weeks_per_year):
-    key = key_prefix + str(week) + "/"
-    file_name = tile_hierarchy.levels[args.level].GetFile(args.tile_id, args.level)
-    key += file_name
-    try:
-      file_path = os.path.dirname(key + ".gz" )
-      if not os.path.exists(directory + file_path):
-        try:
-          os.makedirs(directory + file_path)
-        except OSError as e:
-          if e.errno != errno.EEXIST:
-            raise
-      with open(directory + key + ".gz" , "wb") as f:
-        s3.download_fileobj(args.bucket, key + ".gz", f)
-    except botocore.exceptions.ClientError as e:
-      if e.response['Error']['Code'] != "404":
-        raise
-    decompressedFile = gzip.GzipFile(directory + key + ".gz", mode='rb')
-    with open(directory + key , 'w') as outfile:
-      outfile.write(decompressedFile.read())
+    weeks_per_year = 52
+    week = 0
+    while ( week < weeks_per_year):
+      key = key_prefix + str(week) + "/"
+      file_name = tile_hierarchy.levels[args.level].GetFile(args.tile_id, args.level)
+      key += file_name
+      try:
+        file_path = os.path.dirname(key + ".gz" )
+        if not os.path.exists(directory + file_path):
+          try:
+            os.makedirs(directory + file_path)
+          except OSError as e:
+            if e.errno != errno.EEXIST:
+              raise
+        with open(directory + key + ".gz" , "wb") as f:
+          s3.download_fileobj(args.bucket, key + ".gz", f)
+      except botocore.exceptions.ClientError as e:
+        if e.response['Error']['Code'] != "404":
+          raise
+      decompressedFile = gzip.GzipFile(directory + key + ".gz", mode='rb')
+      with open(directory + key , 'w') as outfile:
+        outfile.write(decompressedFile.read())
 
-    spdFileNames.append(outfile.name)
-    week += 1
-
-################################################################################
+      spdFileNames.append(outfile.name)
+      week += 1
+  ################################################################################
 
   print 'getting avg speeds from list of protobuf speed tile extracts'
-  speedListPerSegment = createAvgSpeedList(spdFileNames)
-  #speedListPerSegment = createAvgSpeedList(args.speedtile_list)
+  if not args.local:
+    log.debug('AWS speed processing...')
+    speedListPerSegment = createAvgSpeedList(spdFileNames)
+  else:
+    log.debug('LOCAL speed processing...')
+    speedListPerSegment = createAvgSpeedList(args.speedtile_list)
   
-  #print 'create reference speed tiles for each segment'
+  print 'create reference speed tiles for each segment'
   createRefSpeedTile(args.ref_tile_path, args.ref_tile_file, speedListPerSegment)
 
   if args.verbose:
